@@ -43,10 +43,7 @@ module ClosureTree
       end
 
       def migration_version
-        major = ActiveRecord::VERSION::MAJOR
-        if major >= 5
-          "[#{major}.#{ActiveRecord::VERSION::MINOR}]"
-        end
+        "[#{ActiveRecord::VERSION::MAJOR}.#{ActiveRecord::VERSION::MINOR}]"
       end
 
       def self.next_migration_number(dirname)
